fixed half of the notes
