+++ src/main/java/com/epam/homelibrary/DAO/UserDAO.java
@@ -1,0 +1,11 @@
+package com.epam.homelibrary.DAO;
+
+import com.epam.homelibrary.models.User;
+
+public interface UserDAO {
+    void createUser(User user);
+
+    void blockUser(String username);
+
+    void getUserLogHistory();
+}
