+++ src/main/resources/configuration.properties
@@ -2,7 +2,3 @@ URL = jdbc:h2:~/test
 USER = sa
 PASSWORD = pwd
 
-
-#<user>sa</user>
-#<password>pwd</password>
-#<url>jdbc:h2:~/test</url>
