+++ src/main/resources/hibernate.cfg.xml
@@ -5,7 +5,6 @@
     <session-factory>
         <property name="hibernate.dialect">org.hibernate.dialect.H2Dialect</property>
         <property name="hibernate.connection.driver_class">org.h2.Driver</property>
-<!--        <property name="hibernate.connection.url">jdbc:h2:D:\Кирилл\EPAM_internship\Module_15\DataBase</property>-->
         <property name="hibernate.connection.url">jdbc:h2:~/test</property>
         <property name="hibernate.connection.username">sa</property>
         <property name="hibernate.connection.password">pwd</property>
