+++ src/main/resources/users.json
@@ -1,1 +1,9 @@
-{"name":"Hermione Granger","login":"User","password":"11111","logHistory":[],"isBlocked":false,"listOfBookmarks":[],"listOfBooks":[]}
+{
+  "name": "Hermione Granger",
+  "login": "User",
+  "password": "11111",
+  "logHistory": [],
+  "isBlocked": false,
+  "listOfBookmarks": [],
+  "listOfBooks": []
+}
